"Remove redundant code and add testing dependencies"
The main change in this commit is the removal of unnecessary lines of code from our Main.java file. We have also made additions to our pom.xml dependencies for testing. The new dependency introduced is "system-rules" whose purpose is to enable tests to interact smoothly with system rules (like System.out).

Moreover, we've introduced a new 'MainTests.java' class for better code maintenance. This test class is specifically created to test the Main class functionalities. We've also expanded the existing tests in 'MovieTests.java' and 'CustomerTests.java' for achieving higher test coverage. These modifications help to enhance the code reliability and maintainability.