Update workspace configuration and recent git actions
The workspace.xml file in .idea directory has been updated. Changes were made both in the ChangeListManager and the entries of recent git branches. The reason for these changes was the creation of a new branch named "jmm/RefactorStatementMethod". New checkout actions and switching between branches led to the update of entries inside the workspace settings, reflecting the most recent activities in the IntelliJ IDEA environment.