Refactor movie charge calculation to Rental class
Moved the responsibility of calculating the movie rental charge from the Customer class to the Rental class.

This is part of an effort to adhere to the Single Responsibility Principle (SRP), by ensuring that each class has only one reason to change. Now, if the logic for the rental charge changes, it only affects the Rental class, not the Customer class.